Fixes to allow bulk put requests with only empty folder objects.
Assume any 2xx status code is expected.  Some requests, like bulk put,
can return either 200 or 204 responses depending on the request.  This
change lets us tolerate that.  In the future, we might want to have a
way to specify multiple expected status codes.

Fix ds3_bulk to tolerate responses with no body, which is the case when
a user is bulk putting only empty folder objects (which results in a 204
response since there's nothing else for the user to do).