Set GLIB2_LIBRARIES to an absolute path when PkgConfig is used.
PkgConfig sets xxx_LIBRARIES to be just the name of the libraries without
the paths or file extensions and sets xxx_LIBRARY_DIRS to be the library
paths, however, target_link_libraries expects an absolute path.  This
uses FIND_LIBRARY after PkgConfig finds the library to get the absolute
path of the library and then override GLIB2_LIBRARIES.

See https://cmake.org/Bug/view.php?id=15804 for a bug that was
filed against CMake regarding this.